Move FlashMessage module
