+++ src/components/cluster_detail/scale_cluster_modal.js
@@ -9,11 +9,7 @@ import { connect } from 'react-redux';
 import { bindActionCreators } from 'redux';
 import cmp from 'semver-compare';
 import * as clusterActions from '../../actions/clusterActions';
-import {
-  FlashMessage,
-  messageType,
-  messageTTL,
-} from '../../actions/flashMessageActions';
+import { FlashMessage, messageType, messageTTL } from '../../lib/flash_message';
 import PropTypes from 'prop-types';
 
 class ScaleClusterModal extends React.Component {
