+++ src/components/create_cluster/release_selector.js
@@ -5,11 +5,7 @@ import cmp from 'semver-compare';
 import Button from '../shared/button';
 import { connect } from 'react-redux';
 import { loadReleases } from '../../actions/releaseActions';
-import {
-  FlashMessage,
-  messageType,
-  messageTTL,
-} from '../../actions/flashMessageActions';
+import { FlashMessage, messageType, messageTTL } from '../../lib/flash_message';
 import _ from 'underscore';
 import ReleaseDetailsModal from '../modals/release_details_modal';
 import PropTypes from 'prop-types';
