+++ src/lib/flash_message.js
